requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cloudinary==1.24.0
cryptography==3.3.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
python-decouple==3.3
python-telegram-bot==13.1
pytz==2020.1
redis==3.5.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
whitenoise==5.2.0
